// src/app/interfaces/event.ts
export interface Event {
  id:               number;
  status:           string;
  date:             string;
  description:      string;
  time:             string;
  start_time_hours: number;
  start_time_min:   number;
  duration:         number;
  customer_id:      number;
  uid;              string;
  customer_name:    string;
}